# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "E - PHARMACY",
    "version": "1.0.0",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of E-PHARMACY"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3001"
    }
  ],
...
  "paths": {
    "/user/register": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Register new User",
        "operationId": "registerUser",
        "description": "Register a new User and create session",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully registered user!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully registered user!"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "user"
                      ],
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/SignUpResponse"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Login User",
        "operationId": "loginUser",
        "description": "Authenticate User and create session",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully logged in a user!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully logged in a user!"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "user"
                      ],
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/SignInResponse"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
...
    "/user/logout": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "User logout",
        "operationId": "logoutUser",
        "description": "Logout User from his session",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Logged out (no content)"
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
...
    "/shop/{shopId}": {
      "get": {
        "tags": [
          "Shop"
        ],
        "summary": "Get Shop by id",
        "operationId": "getShopById",
        "description": "Get Shop by id",
        "parameters": [
          {
            "in": "path",
            "name": "shopId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "68498236a100312bea079030"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found shop with id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found shop with id!"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "shop"
                      ],
                      "properties": {
                        "shop": {
                          "$ref": "#/components/schemas/GetShopIdResponse"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/shop/{shopId}/product/{productId}": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "Get Product by id",
        "operationId": "getProductById",
        "description": "Get Product by id",
        "parameters": [
          {
            "in": "path",
            "name": "shopId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "68498236a100312bea079030"
            }
          },
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "69a5716761228a7ed65a3ae9"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found product with id!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found product with id!"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "product",
                        "reviews"
                      ],
                      "properties": {
                        "product": {
                          "$ref": "#/components/schemas/GetProductByIDResponse"
                        },
                        "reviews": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Review"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/shop/{shopId}/product/{productId}/delete": {
      "delete": {
        "tags": [
          "Product"
        ],
        "summary": "Delete Product",
        "operationId": "deleteProduct",
        "description": "Product Delete",
        "parameters": [
          {
            "in": "path",
            "name": "shopId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "68498236a100312bea079030"
            }
          },
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "69a5716761228a7ed65a3ae9"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Product deleted successfully!"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "SignUpRequest": {
        "type": "object",
        "required": [
          "name",
          "email",
          "phone",
          "password"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "User name",
            "example": "John"
          },
          "email": {
            "type": "string",
            "description": "User email",
            "example": "john@gmail.com"
          },
          "phone": {
            "type": "string",
            "description": "User phone",
            "example": "+380987766555"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "68efd87778"
          }
        }
      },
      "SignUpResponse": {
        "type": "object",
        "required": [
          "name",
          "email",
          "phone",
          "_id",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "User name",
            "example": "John"
          },
          "email": {
            "type": "string",
            "description": "User email",
            "example": "john@gmail.com"
          },
          "phone": {
            "type": "string",
            "description": "User phone",
            "example": "+380987766555"
          },
          "_id": {
            "type": "string",
            "description": "User id",
            "example": "68efd834f1781f960b1e7a6c"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-10-08T18:21:10.157+00:00"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-10-08T18:21:10.157+00:00"
          }
        }
      },
      "SignInRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "description": "Shop email",
            "example": "goods@gmail.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "68efd87778"
          }
        }
      },
      "SignInResponse": {
        "type": "object",
        "required": [
          "accessToken"
        ],
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "User accessToken",
            "example": "TYEJ0iVVw81X8Bn0hX/nykwWhcYQ1EsjAHR"
          }
        }
      },
...
      "GetShopIdResponse": {
        "type": "object",
        "required": [
          "_id",
          "name",
          "email",
          "phone",
          "password",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "User id",
            "example": "68efd834f1781f960b1e7a6c"
          },
          "name": {
            "type": "string",
            "description": "User name",
            "example": "John"
          },
          "email": {
            "type": "string",
            "description": "User email",
            "example": "john@gmail.com"
          },
          "phone": {
            "type": "string",
            "description": "User phone",
            "example": "+380987766555"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "68efd87778"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-10-08T18:21:10.157+00:00"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-10-08T18:21:10.157+00:00"
          }
        }
      },
...
      "GetProductByIDResponse": {
        "type": "object",
        "required": [
          "shopId",
          "id",
          "photo",
          "name",
          "suppliers",
          "stock",
          "price",
          "category",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "shopId": {
            "type": "string",
            "description": "Shop id",
            "example": "68efd834f1781f960b1e7a6c"
          },
          "id": {
            "type": "number",
            "description": "Product number",
            "example": 12
          },
          "photo": {
            "type": "string",
            "description": "Product photo",
            "example": "https://i.ibb.co/X330FTj/shop-4-10-1000x1000-min.jpg"
          },
          "name": {
            "type": "string",
            "description": "Product name",
            "example": "Shampoo"
          },
          "suppliers": {
            "type": "string",
            "description": "Product supplier",
            "enum": [
              "Square",
              "Beximco",
              "Uniliver",
              "ACI",
              "Acme"
            ],
            "example": "Square"
          },
          "stock": {
            "type": "number",
            "description": "Product stock",
            "example": 15
          },
          "price": {
            "type": "number",
            "description": "Product price",
            "example": 45.66
          },
          "category": {
            "type": "string",
            "description": "Product category",
            "enum": [
              "Hand",
              "Head",
              "Medicine",
              "Leg",
              "Dental Care",
              "Heart"
            ],
            "example": "Hand"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-10-08T18:21:10.157+00:00"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-10-08T18:21:10.157+00:00"
          }
        }
      },
      "Review": {
        "type": "object",
        "required": [
          "_id",
          "name",
          "testimonial"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Review id",
            "example": "68efd834f1781f960b1e7a6c"
          },
          "name": {
            "type": "string",
            "description": "Review name",
            "example": "Maria Tkachyk"
          },
          "testimonial": {
            "type": "string",
            "description": "Testimonial result",
            "example": "I recently used this medical platform to book an appointment with a specialist, and I was impressed by how easy and user-friendly the process was. Highly recommended!"
          }
        }
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "BadRequest"
                },
                "data": {
                  "type": "object",
                  "additionalProperties": true,
                  "description": "Error object produced by http-errors (may contain name, message, details, etc.)",
                  "example": {
                    "name": "BadRequest",
                    "message": "Validation failed",
                    "details": [
                      "name is required"
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorized Error"
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "User not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "User not found"
                }
              }
            }
          }
        }
      },
      "409": {
        "description": "Conflict — resource already exists (Email in use)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 409
                },
                "message": {
                  "type": "string",
                  "example": "Conflict"
                },
                "data": {
                  "type": "object",
                  "additionalProperties": true,
                  "description": "Error object produced by http-errors (may contain name, message, details, etc.)",
                  "example": {
                    "name": "Conflict",
                    "message": "Email already in use"
                  }
                }
              }
            }
          }
        }
      },
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    }
  }
}